@@ -1,12 +1,13 @@
 import { Component } from '@angular/core';
 import { TodoService } from './todo.service';
 import { Todo } from './todo';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+  styleUrls: ['./app.component.scss'],
   providers: [TodoService]
 })
 export class AppComponent {
